move `Gui` variants to top level classes
- `Gui.Chest.Builder`, `Gui.Hopper.Builder`, Gui.Dropper.Builder` is verbose
- `Gui.Chest`, `Gui.Hopper`, `Gui.Dropper` is unnatural - it isn't a gui chest, but a chest gui
